new CreateDoc class. Architecture reorganisation
